Use raw string for regex
The \d in this regex is interpreted by python as an escaped unicode
character.  We want it to mean "digit", which would be what \d indicates
in a regex when read as two characters, not one unicode char.  Here we
convert the backslash character to an escape character by prefixing the
string with "r", turning it into a raw strong.